+++ Report_App_WASM/Server/Models/DatabaseConnection.cs
@@ -4,7 +4,7 @@ public class DatabaseConnection : BaseTraceability
 {
     private string? _password;
     private string? _databaseTypeName;
-    public int DatabaseConnectionId { get; set; }
+    public long DatabaseConnectionId { get; set; }
     [MaxLength(20)] public string ConnectionType { get; set; } = "SQL";
     public TypeDb TypeDb { get; set; }
 
@@ -38,7 +38,7 @@ public class DatabaseConnection : BaseTraceability
     public int CommandFetchSize { get; set; } = 131072;
     public bool UseTableMetaData { get; set; } = false;
     public bool UseTableMetaDataFromAnotherProvider { get; set; } = false;
-    public int IdTableMetaData { get; set; }
+    public long IdTableMetaData { get; set; }
     public int AdHocQueriesMaxNbrofRowsFetched { get; set; } = 100000;
     public int TaskSchedulerMaxNbrofRowsFetched { get; set; } = 1000000;
     public int DataTransferMaxNbrofRowsFetched { get; set; } = 2000000;
