+++ Report_App_WASM/Shared/ApiExchanges/TaskActivatePayload.cs
@@ -3,5 +3,5 @@
 public class TaskActivatePayload
 {
     public bool Activate { get; init; }
-    public int ScheduledTaskId { get; init; }
+    public long ScheduledTaskId { get; init; }
 }
