+++ Report_App_WASM/Shared/ApiExchanges/UserConfigurationSave.cs
@@ -5,7 +5,7 @@ public class UserConfigurationSave
     public string? SaveName { get; set; }
     public string? Parameters { get; set; }
     public TypeConfiguration TypeConfiguration { get; set; }
-    public int IdIntConfiguration { get; set; }
+    public long IdIntConfiguration { get; set; }
     public string? IdStringConfiguration { get; set; }
     public string? SavedValues { get; set; }
 }
