+++ Report_App_WASM/Shared/DTO/FileStorageLocationDto.cs
@@ -2,7 +2,7 @@
 
 public class FileStorageLocationDto : BaseTraceabilityDto, IDto
 {
-    [Key] public int FileStorageLocationId { get; set; }
+    [Key] public long FileStorageLocationId { get; set; }
     [Required][MaxLength(250)] public string? ConfigurationName { get; set; }
     [Required][MaxLength(4000)] public string FilePath { get; set; } = ".";
     public bool IsReachable { get; set; }
