+++ Report_App_WASM/Shared/DTO/ScheduledTaskDistributionListDto.cs
@@ -2,7 +2,7 @@
 
 public class ScheduledTaskDistributionListDto : BaseTraceabilityDto, IDto
 {
-    public int ScheduledTaskDistributionListId { get; set; }
+    public long ScheduledTaskDistributionListId { get; set; }
     [MaxLength(4000)] public string Recipients { get; set; } = "[]";
     [MaxLength(4000)] public string? EmailMessage { get; set; }
     public virtual ScheduledTaskDto? ScheduledTask { get; set; }
