+++ Report_App_WASM/Shared/DTO/ScheduledTaskDto.cs
@@ -8,10 +8,10 @@ public sealed class ScheduledTaskDto : BaseTraceabilityDto, IDto
         DistributionLists = new HashSet<ScheduledTaskDistributionListDto>();
     }
 
-    public int ScheduledTaskId { get; set; }
+    public long ScheduledTaskId { get; set; }
     [Required][MaxLength(200)] public string? TaskName { get; set; }
     [Required][MaxLength(250)] public string ProviderName { get; set; }
-    public int IdDataProvider { get; set; }
+    public long IdDataProvider { get; set; }
     [MaxLength(60)] public string? TaskNamePrefix { get; set; }
     public TaskType Type { get; set; }
     [MaxLength(20)] public string? TypeName { get; set; }
@@ -28,7 +28,7 @@ public sealed class ScheduledTaskDto : BaseTraceabilityDto, IDto
     [MaxLength(4000)] public string GlobalQueryParameters { get; set; } = "[]";
     [MaxLength(100)] public string? TimeZone { get; set; } = TimeZoneInfo.Local.Id;
     public DateTime? LastRunDateTime { get; set; } = null;
-    public int FileStorageLocationId { get; set; }
+    public long FileStorageLocationId { get; set; }
     [MaxLength(4000)] public string MiscParameters { get; set; } = "[]";
     public ICollection<ScheduledTaskQueryDto> TaskQueries { get; set; } = new List<ScheduledTaskQueryDto>();
     public ICollection<ScheduledTaskDistributionListDto> DistributionLists { get; set; } = new List<ScheduledTaskDistributionListDto>();
