+++ Report_App_WASM/Shared/DTO/TaskStepLogDto.cs
@@ -2,11 +2,11 @@
 
 public class TaskStepLogDto : IDto
 {
-    public int Id { get; set; }
-    public int TaskLogId { get; set; }
+    public long Id { get; set; }
+    public long TaskLogId { get; set; }
     public DateTime TimeStamp { get; set; } = DateTime.Now;
     [MaxLength(1000)] public string? Step { get; set; }
     public string? Info { get; set; }
     public LogType RelatedLogType { get; set; } = LogType.NotSet;
-    public int RelatedLogId { get; set; }
+    public long RelatedLogId { get; set; }
 }
