+++ Report_App_WASM/Shared/SelectItemDataproviderInfo.cs
@@ -2,11 +2,11 @@
 
 public class SelectItemDataproviderInfo
 {
-    public int DataProviderId { get; init; }
+    public long DataProviderId { get; init; }
     public string? ProviderName { get; init; }
     public string? LogoPath { get; set; }
     public bool HasALogo { get; set; }
     public bool IsVisible { get; init; }
     public bool IsActivated { get; init; }
-    public int DatabaseConnectionId { get; init; }
+    public long DatabaseConnectionId { get; init; }
 }
